+++ example/integration_test/src/_test_imports.dart
@@ -1,0 +1,21 @@
+//
+export 'package:flutter/material.dart';
+
+//
+export 'package:flutter_test/flutter_test.dart';
+
+export 'package:integration_test/integration_test.dart'
+    show IntegrationTestWidgetsFlutterBinding;
+
+// Example app
+export 'package:example/main.dart';
+
+/// Helpful utilities for testing
+export 'test_utils.dart';
+
+/// Individual Tests
+export 'tests/test_page1.dart';
+
+export 'tests/test_page2.dart';
+
+export 'tests/test_page3.dart';