+++ lib/state_set.dart
@@ -1,7 +1,7 @@
 library state_set;
 
 ///
-/// Copyright (C) 2020 Andrious Solutions
+/// Copyright (C) 2024 Andrious Solutions
 ///
 /// Licensed under the Apache License, Version 2.0 (the "License");
 /// you may not use this file except in compliance with the License.
@@ -15,11 +15,39 @@ library state_set;
 /// See the License for the specific language governing permissions and
 /// limitations under the License.
 ///
-///          Created  11 Dec 2020
+///          Created  08 Aug 2024
 ///
 ///
 
-import 'package:flutter/material.dart';
+import 'package:flutter/foundation.dart'
+    show
+        ErrorDescription,
+        FlutterError,
+        FlutterErrorDetails,
+        Key,
+        ValueKey,
+        VoidCallback,
+        kDebugMode,
+        mustCallSuper;
+
+import 'package:flutter/material.dart'
+    show
+        BuildContext,
+        ErrorDescription,
+        FlutterError,
+        FlutterErrorDetails,
+        InheritedElement,
+        InheritedWidget,
+        Key,
+        SizedBox,
+        State,
+        StatefulWidget,
+        StatelessWidget,
+        ValueKey,
+        VoidCallback,
+        Widget,
+        WidgetBuilder,
+        mustCallSuper;
 
 /// Supply the StatefulWidget access to its State object and others
 extension StateMapStatefulWidgetExtension on StatefulWidget {
@@ -51,18 +79,6 @@ extension StateMapStatefulWidgetExtension on StatefulWidget {
 
   /// Call the setState() function from this particular StatefulWidget.
   bool setState(VoidCallback fn) => StateSet.setStateOf(this, fn);
-
-  /// Call the setState() function from this particular StatefulWidget.
-  @Deprecated('Use setState() instead. A recognized Flutter function.')
-  bool refresh() => StateSet.refreshState(this);
-
-  /// Call the setState() function from this particular StatefulWidget.
-  @Deprecated('Use setState() instead. A recognized Flutter function.')
-  bool rebuild() => refresh();
-
-  /// Call the setState() function from this particular StatefulWidget.
-  @Deprecated('Use setState() instead. A recognized Flutter function.')
-  bool notifyListeners() => refresh();
 }
 
 //todo: Does this work?
@@ -115,14 +131,6 @@ mixin StateSet<T extends StatefulWidget> on State<T> {
     super.deactivate();
   }
 
-  /// Remove objects from the static Maps if not already removed.
-  /// add this function to the State object's dispose function
-  @override
-  void dispose() {
-    _removeThisState();
-    super.dispose();
-  }
-
   /// Add the 'current' State entry to the static objects
   void _addThisState() {
     _stateWidgets.addAll({widget.runtimeType: this});
@@ -136,10 +144,6 @@ mixin StateSet<T extends StatefulWidget> on State<T> {
     _statefulStates.removeWhere((key, value) => value == this);
     _setStates.removeWhere((key, value) => value == this);
   }
-
-  /// 'Refresh' the State object's widget tree
-  @Deprecated('Use setState() instead--a recognized Flutter function.')
-  void refresh() => setState(() {});
 
   /// Retrieve the State object by its StatefulWidget
   /// Returns null if not found
@@ -218,169 +222,184 @@ mixin StateSet<T extends StatefulWidget> on State<T> {
     }
     return set;
   }
-
-  /// Call setState() from the State object from the specified StatefulWidget
-  /// Return true if successful
-  /// Usually called in the StatefulWidget passing 'this' as a parameter.
-  static bool refreshState(StatefulWidget? widget) => setStateOf(widget, () {});
-
-  /// Call setState() from the State object from the specified StatefulWidget
-  /// Return true if successful
-  /// Usually called in the StatefulWidget passing 'this' as a parameter.
-  static bool rebuildState(StatefulWidget? widget) => setStateOf(widget, () {});
 }
 
-/// Implement so to serve as a Business Logic Component for a SetState object.
-class StateBLoC<T extends State> {
-  /// Collect the designated State object to work with
-  StateBLoC() {
-    // Note, this is in case State object is available at this time.
-    state = StateSet.to<T>();
-  }
+/// Supplies an InheritedWidget to a State class
+///
+mixin InheritedWidgetStateMixin on State {
+  /// A flag determining whether the built-in InheritedWidget is used or not.
+  bool get useInherited => _useInherited;
+  late bool _useInherited;
 
-  /// The Subclass should supply the appropriate SetState object
-  T? state;
+  // Collect any 'widgets' depending on this State's InheritedWidget.
+  final Set<BuildContext> _dependencies = {};
 
-  /// Call your State object.
-  // ignore: invalid_use_of_protected_member
-  void setState(VoidCallback fn) => state?.setState(fn);
+  InheritedElement? _inheritedElement;
 
-  /// Supply the State object to this BloC object.
-  @mustCallSuper
+  // Widget passed to the InheritedWidget.
+  Widget? _child;
+
+  @override
   void initState() {
-    state = StateSet.to<T>();
+    super.initState();
+    // Supply an identifier to the InheritedWidget
+    _key = ValueKey<State>(this);
   }
 
-  /// Nullify its reference in the State object's own dispose() function.
-  /// Should clean up memory.
-  @mustCallSuper
-  void dispose() {
-    state = null;
-  }
-}
+  Key? _key;
 
-/// Works with an InheritedWidget to allow for targeted and independent
-/// rebuilds throughout your app. A powerful ability, you merely 'link' a
-/// Widget to an InheritedWidget using .dependOnInheritedWidget(context)
-/// and then call that InheritedWidget using .notifyClients() to rebuild its
-/// dependencies.
-///
-/// /// Static API Example
-/// import 'package:state_set/state_set.dart';
-///
-/// /// Static API
-/// class APIImages extends StatelessWidget {
-///   ///
-///   const APIImages({Key? key}) : super(key: key);
-///
-///   static final InheritedWrap _branch = InheritedWrap<_ImageInherited>(
-///     builder: (child) => _ImageInherited(child: child),
-///     child: const HomePage(),
-///   );
-///
-///   /// Link a widget to an InheritedWidget
-///   static bool dependOnInheritedWidget(BuildContext? context) =>
-///       _branch.dependOnInheritedWidget(context);
-///
-///   /// In harmony with Flutter's own API
-///   static void notifyClients() => _branch.notifyClients();
-///
-///   /// Supply a Widget builder
-///   static Widget setState({
-///     Key? key,
-///     required Widget Function(BuildContext context) builder,
-///   }) =>
-///       _SetState(key: key, builder: builder);
-///
-///   /// Determine if
-///   static bool inBuilder = false;
-///
-///   @override
-///   Widget build(BuildContext context) => _branch;
-/// }
-///
-/// /// The InheritedWidget assigned 'dependent' child widgets.
-/// class _ImageInherited extends InheritedWidget {
-///   ///
-///   const _ImageInherited({Key? key, required Widget child})
-///       : super(key: key, child: child);
-///
-///   @override
-///   bool updateShouldNotify(covariant InheritedWidget oldWidget) => true;
-/// }
-///
-/// ///  Used like the function, setState(), to 'spontaneously' call
-/// ///  build() functions here and there in your app. Much like the Scoped
-/// ///  Model's ScopedModelDescendant() class.
-/// ///  This class object will only rebuild if the InheritedWidget notifies it
-/// ///  as it is a dependency.
-/// class _SetState extends StatelessWidget {
-///   /// Supply a 'builder' passing in the App's 'data object' and latest BuildContext object.
-///   const _SetState({Key? key, required this.builder}) : super(key: key);
-///
-///   /// This is called with every rebuild of the App's inherited widget.
-///   final Widget Function(BuildContext context) builder;
-///
-///   /// Calls the required Function object:
-///   /// Function(BuildContext context)
-///   @override
-///   Widget build(BuildContext context) {
-///     /// Go up the widget tree and link to the App's inherited widget.
-///     context.dependOnInheritedWidgetOfExactType<_ImageInherited>();
-///
-///     APIImages.inBuilder = true;
-///
-///     final Widget widget = builder(context);
-///
-///     APIImages.inBuilder = false;
-///
-///     return widget;
-///   }
-/// }
-class InheritedWrap<U extends InheritedWidget> extends StatefulWidget {
+  @override
+  Widget build(BuildContext context) {
+    _buildOverridden = false;
+    if (useInherited) {
+      _child ??= builder(context);
+    } else {
+      _child = builder(context);
+    }
+    return useInherited
+        ? _StateXInheritedWidget(
+            key: _key,
+            mixin: this,
+            child: _child ?? const SizedBox.shrink(),
+          )
+        : _child!;
+  }
+
+  /// A flag. Note if build() function was overridden or not.
+  bool get buildOverridden => _buildOverridden;
+  bool _buildOverridden = true;
+
+  /// Use this function instead of the build() function
   ///
-  const InheritedWrap({
-    Key? key,
-    required this.builder,
-    required this.child,
-  }) : super(key: key);
+  Widget builder(BuildContext context) {
+    _builderOverridden = false;
+    return const SizedBox.shrink();
+  }
 
-  /// Supply a child Widget to the returning InheritedWidget's child parameter.
-  final U Function(Widget child) builder;
+  /// A flag. Note if builder() function was overridden or not.
+  bool get builderOverridden => _builderOverridden;
+  bool _builderOverridden = true;
 
-  /// The 'child' Widget eventually passed to the InheritedWidget.
-  final Widget child;
+  /// Determine if the dependencies should be updated.
+  bool updateShouldNotify(covariant InheritedWidget oldWidget) => true;
 
-  /// Link a widget to a InheritedWidget of type U
+  ///
+  ///  Set the specified widget (through its context) as a dependent of the InheritedWidget
+  ///
+  ///  Return false if not configured to use the InheritedWidget
   bool dependOnInheritedWidget(BuildContext? context) {
-    bool dependOn = context != null;
-    if (dependOn) {
-      final inheritedWidget = context.dependOnInheritedWidgetOfExactType<U>();
-      dependOn = inheritedWidget != null;
+    final depend = useInherited && context != null;
+    if (depend) {
+      if (_inheritedElement == null) {
+        _dependencies.add(context);
+      } else {
+        context.dependOnInheritedElement(_inheritedElement!);
+      }
     }
-    return dependOn;
+    return depend;
   }
 
-  /// Notify any dependencies to be rebuilt.
+  /// In harmony with Flutter's own API there's also a notifyClients() function
+  /// Rebuild the InheritedWidget of the 'closes' InheritedStateX object if any.
   bool notifyClients() {
-    final _state = state;
-    final notified = _state != null;
-    if (notified) {
-      // ignore: invalid_use_of_protected_member
-      _state.setState(() {});
+    final inherited = useInherited;
+    if (inherited) {
+      setState(() {});
     }
-    return notified;
+    return inherited;
   }
 
-  /// Calls setState() function in the Widget's State object.
-  // ignore: invalid_use_of_protected_member
-  void setState(VoidCallback fn) => state?.setState(fn);
+  /// setState() will actually call an InheritedWidget again
+  /// causing a rebuild of any and all dependencies.
+  @override
+  void setState(VoidCallback fn) {
+    //
+    try {
+      super.setState(fn);
+      // catch any errors if called inappropriately
+    } catch (e, stack) {
+      // Throw in DebugMode.
+      if (kDebugMode) {
+        rethrow;
+      } else {
+        //
+        final details = FlutterErrorDetails(
+          exception: e,
+          stack: stack,
+          library: 'state_extended.dart',
+          context: ErrorDescription('setState() error in $this'),
+        );
+        // Resets the count of errors to show a complete error message not an abbreviated one.
+        FlutterError.resetErrorCount();
+        // https://docs.flutter.dev/testing/errors#errors-caught-by-flutter
+        // Log the error.
+        FlutterError.presentError(details);
+      }
+    }
+  }
+
+  /// Called when the State's InheritedWidget is called again
+  /// This 'widget function' will be called again.
+  Widget stateSet(WidgetBuilder? widgetFunc) {
+    widgetFunc ??= (_) => const SizedBox.shrink();
+    return useInherited
+        ? _SetStateXWidget(mixin: this, widgetFunc: widgetFunc)
+        : widgetFunc(context);
+  }
 
   @override
-  State createState() => _InheritedWrapState();
+  void dispose() {
+    _key = null;
+    _child = null;
+    _inheritedElement = null;
+    _dependencies.clear();
+    super.dispose();
+  }
 }
 
-class _InheritedWrapState extends State<InheritedWrap> with StateSet {
+/// The InheritedWidget used by State object
+class _StateXInheritedWidget extends InheritedWidget {
+  const _StateXInheritedWidget({
+    super.key,
+    required this.mixin,
+    required super.child,
+  });
+
+  final InheritedWidgetStateMixin mixin;
+
   @override
-  Widget build(BuildContext context) => widget.builder(widget.child);
+  InheritedElement createElement() {
+    //
+    final element = InheritedElement(this);
+    mixin._inheritedElement = element;
+    // Associate any dependencies widgets to this InheritedWidget
+    // toList(growable: false) prevent concurrent error
+    for (final context in mixin._dependencies.toList(growable: false)) {
+      context.dependOnInheritedElement(element);
+      mixin._dependencies.remove(context);
+    }
+    return element;
+  }
+
+  /// Use the Mixin's updateShouldNotify() function
+  @override
+  bool updateShouldNotify(covariant InheritedWidget oldWidget) =>
+      mixin.updateShouldNotify(oldWidget);
+}
+
+/// Supply a widget to depend upon a State's InheritedWidget
+class _SetStateXWidget extends StatelessWidget {
+  ///
+  const _SetStateXWidget({
+    required this.mixin,
+    required this.widgetFunc,
+  });
+  final InheritedWidgetStateMixin mixin;
+  final WidgetBuilder widgetFunc;
+
+  @override
+  Widget build(BuildContext context) {
+    mixin.dependOnInheritedWidget(context);
+    return widgetFunc(context);
+  }
 }